added functionality to the keyboard
Can now press the equivalent buttons on the keyboard. However, I have yet to figure out how to activate the functions behind the buttons.

The goal would be that they would call the same functions, and serve the same purpose whether you click the button, or press it on your keyboard.